+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 1.3.1 - 25 July 2014
+* Fixed bug in multithread barrier implementation
+* Added Unit test for multithread kernel execution
+
 ### 1.3 - 25 July 2014
 * Restructured project according to F# Project Scaffold
 * Replaced submodules with NuGet package dependencies